+++ src/main/java/christmas/domain/menu/items/Dessert.java
@@ -1,20 +1,19 @@
-package christmas.enums.menuItems;
+package christmas.domain.menu.items;
 
-public enum Beverage implements MenuItem {
-    ZERO_COLA("제로콜라", 3000),
-    RED_WINE("레드와인", 60000),
-    CHAMPAGNE("샴페인", 25000);
+public enum Dessert implements MenuItem {
+    CHOCO_CAKE("초코케이크", 15000),
+    ICE_CREAM("아이스크림", 5000);
 
     private final String name;
     private final int price;
 
-    Beverage(String name, int price) {
+    Dessert(String name, int price) {
         this.name = name;
         this.price = price;
     }
 
     public String getType() {
-        return "Beverage";
+        return "Dessert";
     }
 
     public String getName() {
